requirements.txt
Flask==3.0.0
requests==2.31.0
neo4j
python-dotenv
google-generativeai